+++ tests_corpus/valid/index.sql
@@ -1,0 +1,1 @@
+CREATE INDEX person_major_index ON person(age);